Add support for multiple round of imaging
For now only one workflow is supported.
- Add one set of position (i.e. first round)
- Add same set of position at multiple round

In other words, it does not work with multiple sets of position.

Alignment is done using xcorr of a small region of the image (it is more
reliable). I usually select a couple of nucleus. There is an option to
enter the offset manually. The origin of the image is in the top left
corner.